split out settings sections
