+++ frontend/src/pages/settings/sections/EmailSettings.tsx
@@ -1,0 +1,10 @@
+import type { FC } from 'react'
+
+const EmailSettings: FC = () => (
+  <div className="settings__form">
+    <h2 className="settings__section-heading">Email Settings</h2>
+    <p className="settings__help">TODO</p>
+  </div>
+)
+
+export default EmailSettings
