+++ frontend/src/pages/settings/sections/SmsSettings.tsx
@@ -1,0 +1,10 @@
+import type { FC } from 'react'
+
+const SmsSettings: FC = () => (
+  <div className="settings__form">
+    <h2 className="settings__section-heading">SMS Settings</h2>
+    <p className="settings__help">TODO</p>
+  </div>
+)
+
+export default SmsSettings
